refactor: break Gallery for dif-t components
