+++ culture-portal/src/components/Authors/1/index.tsx
@@ -1,51 +1,19 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import BigPicture, { BigPictureGallery } from "react-bigpicture";
 
-BigPictureGallery.propTypes.noLoader = PropTypes.bool;
-BigPicture.propTypes.noLoader = PropTypes.bool;
-
-const Galery = ({ data }: { data: any }) => {
-  const images = data.gallery.map((el: any) => (
-    <img
-      key={`key${el.fields.file.url}`}
-      src={el.fields.file.url}
-      alt="#"
-      height="200px"
-    />
-
-  ));
-  const videoYoutoobe =
-    <BigPicture
-      type="youtube"
-      src={data.linkVideo}
-    >
-      <img src={data.prevVideo.fields.file.url} alt="#" />
-    </BigPicture>
-    ;
-
-  return (
-    <>
-      <div className='autorsPhoto'><BigPictureGallery loop>{images}</BigPictureGallery></div>
-      {videoYoutoobe}
-    </>
-  )
-};
+import Foto from '../../OverlayFoto';
+import Video from '../../OverlayVideo';
 
 const Author = ({ data }: { data: any }) => {
   console.log(data);
   return (
     <>
       <div> Some info about author {data.nameAuthor}</div>
-      <Galery data={data} />
+      <Foto data={data} />
+      <Video data={data} />
     </>
   )
 };
-
-
-
-
-
 
 Author.propTypes = {
   data: PropTypes.instanceOf(Object).isRequired
